pyproject.toml
[tool.poetry]
name = "azure-text-extractions"
version = "0.1.0"
description = ""
authors = ["Dennis Seah <dennis.seah@microsoft.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
azure-ai-textanalytics = "^5.3.0"
azure-core = "^1.32.0"
azure-identity = "^1.19.0"
python-dotenv = "^1.0.1"
pydantic = "^2.10.4"
lagom = "^2.7.5"
aiohttp = "^3.11.11"
nltk = "^3.9.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
pyright = "^1.1.391"
pre-commit = "^4.0.1"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["typings"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]


[tool.coverage.run]
omit = [
    "*/__init__.py",
    "az_text_extractions/hosting.py",
    "az_text_extractions/protocols/*"
]